Move test factory to testUtils folder
