+++ express-api/tests/unit/controllers/parcels/parcelsController.test.ts
@@ -2,7 +2,7 @@ import { IParcel } from '../../../../controllers/parcels/IParcel';
 import { faker } from '@faker-js/faker';
 import controllers from '../../../../controllers';
 import { Request, Response } from 'express';
-import { MockReq, MockRes, getRequestHandlerMocks } from '../../utilities/factories';
+import { MockReq, MockRes, getRequestHandlerMocks } from '../../../testUtils/factories';
 
 describe('UNIT - Parcels', () => {
   const mockParcel: IParcel = {
